Remove developer mode option to set up cross-signing
